[STREAMPIPES-597] Rename new adapter component to adapter configuration
