+++ ui/src/app/connect/components/adapter-configuration/adapter-configuration-header/adapter-configuration-header.component.html
