+++ ui/src/app/connect/components/adapter-configuration/adapter-configuration.component.html
