+++ ui/src/app/connect/components/adapter-configuration/directives/adapter-template-configuration.directive.ts
@@ -1,34 +1,30 @@
 /*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
+ *  Licensed to the Apache Software Foundation (ASF) under one or more
+ *  contributor license agreements.  See the NOTICE file distributed with
+ *  this work for additional information regarding copyright ownership.
+ *  The ASF licenses this file to You under the Apache License, Version 2.0
+ *  (the "License"); you may not use this file except in compliance with
+ *  the License.  You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
  *
  */
 
 import { Directive, EventEmitter, Input, Output } from '@angular/core';
 import { MatStepper } from '@angular/material/stepper';
-import {
-  AdapterDescriptionUnion,
-  PipelineElementTemplate,
-  PipelineElementTemplateService
-} from '@streampipes/platform-services';
+import { AdapterDescriptionUnion, PipelineElementTemplate, PipelineElementTemplateService } from '@streampipes/platform-services';
 import { FormBuilder } from '@angular/forms';
-import { AdapterTemplateService } from '../../services/adapter-template.service';
+import { AdapterTemplateService } from '../../../services/adapter-template.service';
 import { DialogService } from '@streampipes/shared-ui';
 
 @Directive()
-export abstract class AdapterConfigurationDirective {
+export abstract class AdapterTemplateConfigurationDirective {
 
   /**
    * Adapter description the selected format is added to
