+++ ui/src/app/connect/components/adapter-configuration/format-configuration/format-configuration.component.ts
