+++ ui/src/app/connect/components/adapter-configuration/format-configuration/format-item-json/format-item-json.component.scss
