+++ ui/src/app/connect/components/adapter-configuration/format-configuration/format-item/format-item.component.scss
