+++ ui/src/app/connect/components/adapter-configuration/format-configuration/format-list/format-list.component.ts
