+++ ui/src/app/connect/components/adapter-configuration/generic-adapter-configuration/generic-adapter-configuration.component.html
