+++ ui/src/app/connect/components/adapter-configuration/schema-editor/error-message/error-message.component.ts
