+++ ui/src/app/connect/components/adapter-configuration/schema-editor/event-property-row/event-property-row.component.scss
