+++ ui/src/app/connect/components/adapter-configuration/schema-editor/event-schema-preview/event-schema-preview.component.ts
