+++ ui/src/app/connect/components/adapter-configuration/schema-editor/event-schema/event-schema.component.ts
