+++ ui/src/app/connect/components/adapter-configuration/schema-editor/loading-message/loading-message.component.html
