+++ ui/src/app/connect/components/adapter-configuration/schema-editor/schema-editor-header/schema-editor-header.component.scss
