+++ ui/src/app/connect/components/adapter-configuration/specific-adapter-configuration/specific-adapter-configuration.component.html
