+++ ui/src/app/connect/components/adapter-configuration/start-adapter-configuration/start-adapter-configuration.component.html
