+++ ui/src/app/connect/connect.module.ts
@@ -24,8 +24,8 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatGridListModule } from '@angular/material/grid-list';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 
-import { NewAdapterComponent } from './components/new-adapter/new-adapter.component';
-import { EventSchemaComponent } from './components/new-adapter/schema-editor/event-schema/event-schema.component';
+import { AdapterConfigurationComponent } from './components/adapter-configuration/adapter-configuration.component';
+import { EventSchemaComponent } from './components/adapter-configuration/schema-editor/event-schema/event-schema.component';
 
 import { CustomMaterialModule } from '../CustomMaterial/custom-material.module';
 
@@ -39,8 +39,8 @@ import { TransformationRuleService } from './services/transformation-rule.servic
 import { ConnectService } from './services/connect.service';
 import { AdapterDescriptionComponent } from './components/data-marketplace/adapter-description/adapter-description.component';
 import { DataMarketplaceComponent } from './components/data-marketplace/data-marketplace.component';
-import { FormatItemComponent } from './components/new-adapter/format-configuration/format-item/format-item.component';
-import { FormatListComponent } from './components/new-adapter/format-configuration/format-list/format-list.component';
+import { FormatItemComponent } from './components/adapter-configuration/format-configuration/format-item/format-item.component';
+import { FormatListComponent } from './components/adapter-configuration/format-configuration/format-list/format-list.component';
 import { IconService } from './services/icon.service';
 import { UnitProviderService } from './services/unit-provider.service';
 
@@ -64,8 +64,8 @@ import {
 } from './dialog/edit-event-property/components/edit-unit-transformation/edit-unit-transformation.component';
 import { EditEventPropertyComponent } from './dialog/edit-event-property/edit-event-property.component';
 import { PipelineElementRuntimeInfoComponent } from './components/runtime-info/pipeline-element-runtime-info.component';
-import { EventPropertyRowComponent } from './components/new-adapter/schema-editor/event-property-row/event-property-row.component';
-import { EventSchemaPreviewComponent } from './components/new-adapter/schema-editor/event-schema-preview/event-schema-preview.component';
+import { EventPropertyRowComponent } from './components/adapter-configuration/schema-editor/event-property-row/event-property-row.component';
+import { EventSchemaPreviewComponent } from './components/adapter-configuration/schema-editor/event-schema-preview/event-schema-preview.component';
 import { CoreUiModule } from '../core-ui/core-ui.module';
 import {
   EditCorrectionValueComponent
@@ -73,21 +73,21 @@ import {
 import { ExistingAdaptersComponent } from './components/existing-adapters/existing-adapters.component';
 import {
   SpecificAdapterConfigurationComponent
-} from './components/new-adapter/specific-adapter-configuration/specific-adapter-configuration.component';
+} from './components/adapter-configuration/specific-adapter-configuration/specific-adapter-configuration.component';
 import { ConfigurationGroupComponent } from './components/configuration-group/configuration-group.component';
-import { FormatConfigurationComponent } from './components/new-adapter/format-configuration/format-configuration.component';
+import { FormatConfigurationComponent } from './components/adapter-configuration/format-configuration/format-configuration.component';
 import {
   GenericAdapterConfigurationComponent
-} from './components/new-adapter/generic-adapter-configuration/generic-adapter-configuration.component';
-import { ErrorMessageComponent } from './components/new-adapter/schema-editor/error-message/error-message.component';
-import { LoadingMessageComponent } from './components/new-adapter/schema-editor/loading-message/loading-message.component';
-import { SchemaEditorHeaderComponent } from './components/new-adapter/schema-editor/schema-editor-header/schema-editor-header.component';
+} from './components/adapter-configuration/generic-adapter-configuration/generic-adapter-configuration.component';
+import { ErrorMessageComponent } from './components/adapter-configuration/schema-editor/error-message/error-message.component';
+import { LoadingMessageComponent } from './components/adapter-configuration/schema-editor/loading-message/loading-message.component';
+import { SchemaEditorHeaderComponent } from './components/adapter-configuration/schema-editor/schema-editor-header/schema-editor-header.component';
 import {
   StartAdapterConfigurationComponent
-} from './components/new-adapter/start-adapter-configuration/start-adapter-configuration.component';
+} from './components/adapter-configuration/start-adapter-configuration/start-adapter-configuration.component';
 import { DeleteAdapterDialogComponent } from './dialog/delete-adapter-dialog/delete-adapter-dialog.component';
 import { PlatformServicesModule } from '@streampipes/platform-services';
-import { FormatItemJsonComponent } from './components/new-adapter/format-configuration/format-item-json/format-item-json.component';
+import { FormatItemJsonComponent } from './components/adapter-configuration/format-configuration/format-item-json/format-item-json.component';
 import { RouterModule } from '@angular/router';
 import { SharedUiModule } from '@streampipes/shared-ui';
 import { SpConnectFilterToolbarComponent } from './components/filter-toolbar/filter-toolbar.component';
@@ -100,11 +100,11 @@ import {
 import { SpEpSettingsSectionComponent } from './dialog/edit-event-property/components/ep-settings-section/ep-settings-section.component';
 import {
   SpAdapterOptionsPanelComponent
-} from './components/new-adapter/start-adapter-configuration/adapter-options-panel/adapter-options-panel.component';
+} from './components/adapter-configuration/start-adapter-configuration/adapter-options-panel/adapter-options-panel.component';
 import { SpAdapterTemplateDialogComponent } from './dialog/adapter-template/adapter-template-dialog.component';
 import { JsonPrettyPrintPipe } from './filter/json-pretty-print.pipe';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { AdapterConfigurationHeaderComponent } from './components/new-adapter/adapter-configuration-header/adapter-configuration-header.component';
+import { AdapterConfigurationHeaderComponent } from './components/adapter-configuration/adapter-configuration-header/adapter-configuration-header.component';
 
 @NgModule({
   imports: [
@@ -139,11 +139,11 @@ import { AdapterConfigurationHeaderComponent } from './components/new-adapter/ad
           },
           {
             path: 'create/:appId',
-            component: NewAdapterComponent,
+            component: AdapterConfigurationComponent,
           },
           {
             path: 'edit/:appId',
-            component: NewAdapterComponent,
+            component: AdapterConfigurationComponent,
           }]
       }]),
     SharedUiModule
@@ -173,7 +173,7 @@ import { AdapterConfigurationHeaderComponent } from './components/new-adapter/ad
     FormatItemComponent,
     FormatListComponent,
     JsonPrettyPrintPipe,
-    NewAdapterComponent,
+    AdapterConfigurationComponent,
     SpAdapterTemplateDialogComponent,
     PipelineElementRuntimeInfoComponent,
     TimestampPipe,
